requirements.txt
streamlit
pillow
pytesseract
pdf2image
python-docx
sentence-transformers
faiss-cpu